+++ lms-hibernate/src/main/java/com/ss/lmshibernate/repository/AuthorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.ss.lmshibernate.entity.Author;
 
 @Repository
-public interface AuthorRepository extends JpaRepository<Author, Long> {
+public interface AuthorRepository extends JpaRepository<Author, Integer> {
 
 }
